Refactor 'create' subcommand.
